more splitting and movinf xml params
