+++ waveform_editor/shape_editor/xml_param/direct_param.xml
