+++ waveform_editor/shape_editor/xml_param/inverse_param.xml
